Move custody-probe into own repo
This will simplify the versioning i.e. we can use Git tags will be for
custody-probe alone, as well as the discoverability of custody-probe since
it’ll be indexed separately in GitHub search etc and custody’s README can
link to it, and custody-probe didn’t end up sharing any code with custody.