Optionally show desktop notifications when processes’ states change
Clicking the "Show" button on the notification will activate the Terminal
and open custody to the corresponding service's logs.

This is optional and must be enabled by passing the `—notifications` flag to
`custody-cli` in case it’s annoying in its current form. We should collect UX
feedback and refine the notification preferences, debouncing, etc. before
making it the default.

Fixes https://github.com/mixmaxhq/custody/issues/7.