Add navigational shortcut to process controls
We already handled Escape to close the process details, but now we tell the
user that in case they’re confused.

A cool side effect of refactoring the controls to be derived from props is that
the stop/start toggle’s label will update to show which one is selected.

Refs https://github.com/mixmaxhq/custody/issues/39.