Make it possible to restart processes in groups
The direct fix is in passing `Process#daemonName` to the supervisor client
rather than `Process#name`. But rather than further complicate the ad-hoc
definition of the process model in `ProcessMonitor` (with that change), the
model has been factored into its own class.

Now that processes are class instances and not just “bags of properties”,
`ProcessTable` has been refactored to avoid stripping the processes of those
properties and creating new intermediate process objects in state, but rather
filtering the processes on-demand as is described here
https://reactjs.org/blog/2018/06/07/you-probably-dont-need-derived-state.html#what-about-memoization.

The deep-clone hack in `Console` is no longer necessary, not sure why.

Fixes https://github.com/mixmaxhq/custody/issues/62.