Restart custody when it comes close to running out of memory
Where “close” is currently defined as 80% of the heap size limit, checked
every 30 seconds.

This is a workaround for https://github.com/mixmaxhq/custody/issues/48
but also maybe a good practice for a process that runs indefinitely given the
risk of _some_ sort of crash or memory leak over that time.

To make the restart experience more seamless, we now track the last-selected
process on disk, and restore it after a crash. We could restore the
last-selected process on every launch but if the user explicitly shuts custody
down I think it’s nice for it to start in a clean state. We can revisit that later if
we want.