Add support for displaying maintail output
This output is now an alternate view to the process table / process details
view hierarchy. That is, in either of the two latter views, the user can
press 'tab' to show the maintail output and then again to return to the
previous view.

Fixes https://github.com/mixmaxhq/custody/issues/57.