Remove GitHub dependencies
For reasons similar to https://github.com/clearbit/clearbit-node/issues/43.

We have to update `react-blessed` as well because that loads `blessed` as a
peer dependency. We don’t have to fork the library to load
`@mixmaxhq/blessed` as a peer dependency instead, luckily, since a more
recent version of `react-blessed` adds support for injecting a `blessed`
instance. It appears that this more recent version and/or the use of
`createBlessedRenderer` fixes the issue where `react-blessed` automatically
suppresses the project exiting, too (see that we are now able to import the
library outside `start` without e.g. preventing `custody-cli -h` from exiting).

Fixes https://mixmaxhq.atlassian.net/browse/MX-7481.